[REFACTOR] all code base was changed to my own interpretation of clean architecture based on SOLID and MVVM pattern
