+++ lib/features/world_timezone/data/datasource/remote/Remote.dart
@@ -1,5 +1,5 @@
 import 'package:timezone_locations_app/features/world_timezone/domain/model/world_timezone_model.dart';
 
-abstract class IRepository {
+abstract class IRemote {
   Future<WorldTimezone> getTimezoneInfo(String timezone);
 }
