+++ lib/features/world_timezone/data/datasource/remote/remote_http_impl.dart
@@ -1,11 +1,11 @@
 import 'dart:convert';
 
 import 'package:http/http.dart';
-import 'package:timezone_locations_app/features/world_timezone/data/datasource/remote/model/world_timezone_model.dart';
-import 'package:timezone_locations_app/features/world_timezone/data/datasource/remote_datasource.dart';
+import 'package:timezone_locations_app/features/world_timezone/data/datasource/remote/Remote.dart';
+import 'package:timezone_locations_app/features/world_timezone/data/model/world_timezone_model.dart';
 import 'package:timezone_locations_app/features/world_timezone/domain/model/world_timezone_model.dart';
 
-class RemoteDatasourceImpl extends RemoteDatasource {
+class HttpRemote extends IRemote {
   static final String _endPoint = 'https://worldtimeapi.org/api/timezone/';
   @override
   Future<WorldTimezone> getTimezoneInfo(String timezone) async {
