+++ lib/features/world_timezone/di/DI.dart
@@ -1,9 +1,9 @@
-import 'package:timezone_locations_app/features/world_timezone/data/datasource/remote/remote_datasource_impl.dart';
-import 'package:timezone_locations_app/features/world_timezone/data/repository/common_repository.dart';
+import 'package:timezone_locations_app/features/world_timezone/data/datasource/remote/remote_http_impl.dart';
+import 'package:timezone_locations_app/features/world_timezone/data/repository/repository_impl.dart';
 import 'package:timezone_locations_app/features/world_timezone/domain/repository/repository.dart';
 
 class DI {
-  static final Repository repository = CommonRepository(
-    RemoteDatasourceImpl(),
+  static final IRepository repository = RepositoryImpl(
+    HttpRemote(),
   );
 }
